requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
torch              # matching whisper
google-genai       # Google Gen AI Python SDK (pip package)
python-multipart
google-adk
